Se agrego pagina con tabla de servicios, y administrar salones y servicios por parte del administrador
