+++ admin.html
@@ -41,83 +41,23 @@
           <li class="nav-item">
             <a class="nav-link nav-custom active" href="admin.html">Admin</a>
           </li>
+            </ul>
+          </li>
         </ul>
       </div>
     </div>
   </nav>
 </header>
 
-<div class="container py-4">
-
-  <div class="d-flex justify-content-end mb-3">
-    <a href="usuarios.html" class="btn btn-primary">Ir a Usuarios</a>
+<div class="container text-center mt-5">
+  <h1>Administración</h1>
+  <p>Seleccione que desea administrar:</p>
+  
+  <div class="d-flex justify-content-center gap-3">
+    <a href="adminSalones.html" class="btn btn-primary btn-lg">Administrar Salones</a>
+    <a href="adminServicios.html" class="btn btn-secondary btn-lg">Administrar Servicios</a>
   </div>
-
-  <h1>Administrar Salones</h1>
-
-  <table class="table table-bordered table-striped border-black caption-top table-responsive">
-    <caption>
-      <div class="d-flex flex-row justify-content-between align-items-center">
-        <p> Lista de Salones </p>
-        <a href="crear.html" class="btn btn-success">
-          + Nuevo Salón
-        </a>
-      </div>
-    </caption>
-    <thead>
-      <tr class="table-info border-black">
-        <th>Nombre</th>
-        <th>Descripción</th>
-        <th>Dirección</th>
-        <th>Capacidad</th>
-        <th>Valor</th>
-        <th>Estado</th>
-        <th>Imagen</th>
-        <th>Acciones</th>
-      </tr>
-    </thead>
-    <tbody id="tablaSalonesBody" class="table-group-divider"></tbody>
-  </table>
-
-  <form id="formSalon" class="mb-4">
-    <input type="hidden" id="salonId" />
-    <div class="mb-3">
-      <label for="nombre" class="form-label">Nombre</label>
-      <input type="text" id="nombre" class="form-control" required />
-    </div>
-    <div class="mb-3">
-      <label for="descripcion" class="form-label">Descripción</label>
-      <textarea id="descripcion" class="form-control" required></textarea>
-    </div>
-    <div class="mb-3">
-      <label for="direccion" class="form-label">Dirección</label>
-      <input type="text" id="direccion" class="form-control" required />
-    </div>
-    <div class="mb-3">
-      <label for="capacidad" class="form-label">Capacidad</label>
-      <input type="number" id="capacidad" class="form-control" required min="1" />
-    </div>
-    <div class="mb-3">
-      <label for="valor" class="form-label">Valor</label>
-      <input type="number" id="valor" class="form-control" required min="0" />
-    </div>
-    <div class="mb-3">
-      <label for="estado" class="form-label">Estado</label>
-      <select id="estado" class="form-select" required>
-        <option value="Disponible">Disponible</option>
-        <option value="No disponible">No disponible</option>
-      </select>
-    </div>
-    <div class="mb-3">
-      <label for="imagen" class="form-label">Nombre base de la imagen</label>
-      <input type="text" id="imagen" class="form-control" required />
-    </div>
-    <button type="submit" class="btn btn-primary">Guardar</button>
-    <button type="button" id="btnCancelar" class="btn btn-secondary">Cancelar</button>
-  </form>
 </div>
-
-<script type="module" src="./js/admin.js"></script>
 
 </body>
 </html>
