+++ adminSalones.html
@@ -1,0 +1,123 @@
+<!DOCTYPE html>
+<html lang="es">
+<head>
+  <meta charset="UTF-8" />
+  <title>Administración de Salones</title>
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" />
+  <link href="https://fonts.googleapis.com/css2?family=Fredoka:wght@400;600&display=swap" rel="stylesheet" />
+  <link rel="stylesheet" href="css/styles.css" />
+</head>
+<body>
+
+<script>
+  const token = sessionStorage.getItem('accessToken');
+  if (!token) {
+    window.location.href = 'login.html';
+  }
+</script>
+
+<header>
+  <nav class="navbar navbar-expand-lg navbar-dark bg-gradient shadow-sm">
+    <div class="container-fluid">
+      <a class="navbar-brand fw-bold fs-4" href="index.html">IDW S.A</a>
+      <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav"
+        aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+        <span class="navbar-toggler-icon"></span>
+      </button>
+      <div class="collapse navbar-collapse justify-content-lg-end justify-content-center" id="navbarNav">
+        <ul class="navbar-nav text-center text-lg-end">
+          <li class="nav-item">
+            <a class="nav-link nav-custom" href="index.html">Inicio</a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link nav-custom" href="info.html">Institucional</a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link nav-custom" href="contacto.html">Contacto</a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link nav-custom" href="servicios.html">Servicios</a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link nav-custom active" href="admin.html">Admin</a>
+          </li>
+        </ul>
+      </div>
+    </div>
+  </nav>
+</header>
+
+<div class="container py-4">
+
+  <div class="d-flex justify-content-end mb-3">
+    <a href="usuarios.html" class="btn btn-primary">Ir a Usuarios</a>
+  </div>
+
+  <h1>Administrar Salones</h1>
+
+  <table class="table table-bordered table-striped border-black caption-top table-responsive">
+    <caption>
+      <div class="d-flex flex-row justify-content-between align-items-center">
+        <p> Lista de Salones </p>
+        <a href="crear.html" class="btn btn-success">
+          + Nuevo Salón
+        </a>
+      </div>
+    </caption>
+    <thead>
+      <tr class="table-info border-black">
+        <th>Nombre</th>
+        <th>Descripción</th>
+        <th>Dirección</th>
+        <th>Capacidad</th>
+        <th>Valor</th>
+        <th>Estado</th>
+        <th>Imagen</th>
+        <th>Acciones</th>
+      </tr>
+    </thead>
+    <tbody id="tablaSalonesBody" class="table-group-divider"></tbody>
+  </table>
+
+  <form id="formSalon" class="mb-4">
+    <input type="hidden" id="salonId" />
+    <div class="mb-3">
+      <label for="nombre" class="form-label">Nombre</label>
+      <input type="text" id="nombre" class="form-control" required />
+    </div>
+    <div class="mb-3">
+      <label for="descripcion" class="form-label">Descripción</label>
+      <textarea id="descripcion" class="form-control" required></textarea>
+    </div>
+    <div class="mb-3">
+      <label for="direccion" class="form-label">Dirección</label>
+      <input type="text" id="direccion" class="form-control" required />
+    </div>
+    <div class="mb-3">
+      <label for="capacidad" class="form-label">Capacidad</label>
+      <input type="number" id="capacidad" class="form-control" required min="1" />
+    </div>
+    <div class="mb-3">
+      <label for="valor" class="form-label">Valor</label>
+      <input type="number" id="valor" class="form-control" required min="0" />
+    </div>
+    <div class="mb-3">
+      <label for="estado" class="form-label">Estado</label>
+      <select id="estado" class="form-select" required>
+        <option value="Disponible">Disponible</option>
+        <option value="No disponible">No disponible</option>
+      </select>
+    </div>
+    <div class="mb-3">
+      <label for="imagen" class="form-label">Nombre base de la imagen</label>
+      <input type="text" id="imagen" class="form-control" required />
+    </div>
+    <button type="submit" class="btn btn-primary">Guardar</button>
+    <button type="button" id="btnCancelar" class="btn btn-secondary">Cancelar</button>
+  </form>
+</div>
+
+<script type="module" src="./js/admin.js"></script>
+
+</body>
+</html>
