+++ adminServicios.html
@@ -2,13 +2,21 @@
 <html lang="es">
 <head>
   <meta charset="UTF-8" />
-  <title> Servicios</title>
+  <title>Administración de Servicios</title>
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" />
   <link href="https://fonts.googleapis.com/css2?family=Fredoka:wght@400;600&display=swap" rel="stylesheet" />
   <link rel="stylesheet" href="css/styles.css" />
 </head>
-<body class="d-flex flex-column min-vh-100">
+<body>
+
+<script>
+  const token = sessionStorage.getItem('accessToken');
+  if (!token) {
+    window.location.href = 'login.html';
+  }
+</script>
+
 
   <header>
     <nav class="navbar navbar-expand-lg navbar-dark bg-gradient shadow-sm">
@@ -32,11 +40,12 @@
   </header>
 
   <main class="container py-4">
-    <h1>Servicios disponibles</h1>
+    <h1>Administrar Servicios</h1>
     <table class="table table-bordered table-striped border-black caption-top table-responsive">
       <caption>
         <div class="d-flex flex-row justify-content-between align-items-center">
           <p> Lista de Servicios </p>
+          <a href="#" id="btnNuevoServicio" class="btn btn-success">+ Nuevo Servicio</a>
         </div>
       </caption>
       <thead>
@@ -45,23 +54,39 @@
           <th>Descripción</th>
           <th>Precio</th>
           <th>Estado</th>
+          <th>Acciones</th>
         </tr>
       </thead>
       <tbody id="tablaServiciosBody" class="table-group-divider"></tbody>
     </table>
 
+    <form id="formServicio" class="mb-4 d-none">
+      <input type="hidden" id="servicioId" />
+      <div class="mb-3">
+        <label for="nombre" class="form-label">Nombre</label>
+        <input type="text" id="nombre" class="form-control" required />
+      </div>
+      <div class="mb-3">
+        <label for="descripcion" class="form-label">Descripción</label>
+        <textarea id="descripcion" class="form-control" required></textarea>
+      </div>
+      <div class="mb-3">
+        <label for="precio" class="form-label">Precio</label>
+        <input type="number" id="precio" class="form-control" min="0" required />
+      </div>
+      <div class="mb-3">
+        <label for="estado" class="form-label">Estado</label>
+        <select id="estado" class="form-select" required>
+          <option value="Disponible">Disponible</option>
+          <option value="No disponible">No disponible</option>
+        </select>
+      </div>
+      <button type="submit" class="btn btn-primary">Guardar</button>
+      <button type="button" id="btnCancelar" class="btn btn-secondary">Cancelar</button>
+    </form>
   </main>
 
-  <footer class="bg-dark text-white text-center py-3">
-    <p class="mb-0">
-      &copy; <span id="currentYear"></span> IDW S.A - Todos los derechos
-      reservados
-    </p>
-  </footer>
-
-  <script type="module" src="js/servicios.js"></script>
-  <script>
-    document.getElementById('currentYear').textContent = new Date().getFullYear();
-  </script>
+  <script type="module" src="js/adminServicios.js"></script>
+  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
